+++ src/core/bytebuff.h
@@ -3,9 +3,6 @@
 
 #include "constants.h"
 #include "logger.h"
-#include <type_traits>
-
-#define is(A,B) (std::is_same<A,B>::value)
 
 class bytebuff {
 private:
@@ -39,62 +36,13 @@ public:
     u64 len();
 
     vec<u8> subbuff(u64 pos, u32 len);
-
-    template<class T>
-    T get(){
-        if(is(T,u8) || is(T,u16) ||is(T,u32) ||is(T,u64) ||
-           is(T,i8) || is(T,i16) ||is(T,i32) ||is(T,i64)) {
-            T value = 0;
-            for (u32 i = 0; i < sizeof(T); i++) {
-                value |= (T) data[pos] << (8 * i);
-                pos++;
-            }
-            return value;
-        }else{
-            Log::e() << "Unsupported type" << typeid(T).name() << endl();
-            return (T)null;
-        }
-    }
-
-    template<class T>
-    void put(T item){
-        if(is(T,u8) || is(T,u16) ||is(T,u32) ||is(T,u64) ||
-           is(T,i8) || is(T,i16) ||is(T,i32) ||is(T,i64)) {
-            u32 len = sizeof(T);
-            while (len > 0) {
-                u8 t = (u8) item;
-                data.push_back(t);
-                item = item >> 8;
-                len--;
-            }
-        }else{
-            Log::e() << "Unsupported type" << typeid(T).name() << endl();
-        }
-    }
-
-    template<class T>
-    vec<T> getVec(){
-        u32 size = get<u32>();
-        vec<T> items;
-        for(u32 i = 0; i < size;i++){
-            T temp = get<T>();
-            items.push_back(temp);
-        }
-        return items;
-    }
-
-    template<class T>
-    void putVec(const vec<T>& items){
-        put<u32>(items.size());
-        for(T item : items){
-            put<T>(item);
-        }
-    }
-
-
     void load(const str&);
     void load(const vec<u8>&);
 
+    template<class T> T get();
+    template<class T> void put(T);
+    template<class T> vec<T> getVec();
+    template<class T> void putVec(const vec<T>&);
 };
 
 #endif //FCOIN_BYTEBUFF_H